requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
praw==4.5.1
prawcore==0.10.1
PyYAML==3.12
requests==2.17.3
update-checker==0.16
urllib3==1.21.1
Werkzeug==0.12.2
WTForms==2.1
